hazelcast-guides/hazelcast-platform-operator-expose-externally/docs/modules/ROOT/examples/operator-expose-externally/python-unisocket/requirements.txt
hazelcast-python-client >= 5.0
